+++ changelog.md
@@ -1,4 +1,8 @@
 # Changelogs:
+## v1.2.0:
+### - added the ability to copy records summary
+### - added the rule book
+### - pages loading speed is even faster
 ## v1.1.0:
 ### - boosted pages loading speed
 ### - re-design the `top-players` page